requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Pillow==10.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
